+++ src/mako/database/exceptions/DatabaseException.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\database\exceptions;
+
+use RuntimeException;
+
+/**
+ * Database exception.
+ */
+class DatabaseException extends RuntimeException
+{
+
+}