+++ src/mako/database/exceptions/NotFoundException.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\database\exceptions;
+
+/**
+ * Not found exception.
+ */
+class NotFoundException extends DatabaseException
+{
+
+}